+++ packages/kitchensink/service-fetchq/feature-home.js
@@ -1,0 +1,11 @@
+const homeRouteHandler = async (request) => {
+  // const jwt = request.getContext('jwt');
+  // return jwt.sign({ foo: 'bar ' });
+  return 'fetchq';
+};
+
+module.exports = {
+  name: 'home',
+  hook: '$FASTIFY_GET',
+  handler: [{ url: '/', handler: homeRouteHandler }],
+};