+++ packages/kitchensink/service-fetchq/service-fetchq.test.js
@@ -1,8 +1,8 @@
-describe('service-hash', () => {
+describe('service-fetchq', () => {
   it('Should return a signed object', async () => {
     const res = await get('/');
 
     // A valid JWT should have 3 dots in it
-    expect(res.split('.').length).toBe(2);
+    expect(res).toBe('fetchq');
   });
 });