+++ CHANGELOG.md
@@ -25,6 +25,8 @@
   - Remove Babel
 - service-hash
   - Remove Babel
+- service-fetchq
+  - Remove Babel
 - service-fastify
   - Fixes `$REGISTER_ROUTE` hook so to be synchronous and report injections in the boot report
   - Remove babel and the build step
@@ -36,6 +38,7 @@
   - Added examples for `@forrestjs/service-fastify`
   - Added examples for `@forrestjs/service-jwt`
   - Added examples for `@forrestjs/service-hash`
+  - Added examples for `@forrestjs/service-fetchq`
   - Added `app01` examples for a complex app
 
 ## 3.20.2