+++ packages/kitchensink/CHANGELOG.md
@@ -6,4 +6,5 @@
 - Added examples for `@forrestjs/service-fastify`
 - Added examples for `@forrestjs/service-jwt`
 - Added examples for `@forrestjs/service-hash`
+- Added examples for `@forrestjs/service-fetchq`
 - Added `app01` examples for a complex app