+++ packages/service-fetchq/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## v3.21.0
+
+- Remove Babel
+
 ## v3.13.6
 
 hook FETCHQ_REGISTER_WORKER can return an array of definitions